Support collections of primitives.
This required an API change, but one that was already due. Audit metadata
for collection entries is held by two base classes (keyed collections vs.
sets) and the 'value' of a collection entry is always a single property,
mapped as a component type if necessary. This more closely mirrors how
NHibernate's collection persistence works internally, while also
simplifying some of the audit code which no longer needs to be told about
property names.

This also reduces boilerplate in the mappings.