Collections of entities are now supported.
I'm not sure how much sense it makes to audit anything except primary key
references for collections of entities, therefore trying to audit a
collection which uses property-ref is treated as an error.

Auditing mapped value types is no longer supported in collections because
I don't know if we need it and it was hard to retain. Reimplement it if
necessary.