Expose capture of audit snapshots.
When identifying the results of an operation in the history model, it is
not sufficient to simply assume that the clock can be asked for 'now'
because the audit library may truncate the date, or deferring flush.

For safety, expose the ability to force an immediate flush and capture the
exact datestamp used to audit the operation.