Write new entity history rows for collection changes.
Modifying a child collection of a versioned entity increments that
entity's version. In order to retain integrity of history, it is necessary
to record this change in the audit table.

It is not necessary to retain these rows forever (they could be compacted)
but every previousVersionId must alwasy reference a valid versionId.